Crate datafusion_common

Re-exports

pub use parsers::parse_interval;

pub use scalar::ScalarType;

pub use scalar::ScalarValue;

pub use stats::ColumnStatistics;

pub use stats::Statistics;

pub use stats::TermMeta;

Modules

cast

This module provides DataFusion specific casting functions that provide error handling. They are intended to “never fail” but provide an error message rather than a panic, as the corresponding kernels in arrow-rs such as as_boolean_array do.

config

DataFusion Configuration Options

delta

Contains utility functions for shifting Date objects.

from_slice

A trait to define from_slice functions for arrow types

parsers

Interval parsing logic

scalar

This module provides ScalarValue, an enum that can be used for storage of single elements

stats

This module provides data structures to represent statistics

test_util

Utility functions to make testing DataFusion based crates easier

utils

This module provides the bisect function, which implements binary search.

Macros

assert_contains

A macro to assert that one string is contained within another with a nice error message if they are not.

assert_not_contains

A macro to assert that one string is NOT contained within another with a nice error message if they are are.

context

downcast_value

Downcast an Arrow Array to a concrete type, return an DataFusionError::Internal if the cast is not possible. In normal usage of DataFusion the downcast should always succeed.

internal_err

plan_err

unwrap_or_internal_err

Unwrap an Option if possible. Otherwise return an DataFusionError::Internal. In normal usage of DataFusion the unwrap should always succeed.

Structs

Column

A named reference to a qualified field in a schema.

DFField

DFField wraps an Arrow field and adds an optional qualifier

DFSchema

DFSchema wraps an Arrow schema and adds relation names

ResolvedTableReference

A resolved path to a table of the form “catalog.schema.table”

Enums

DataFusionError

DataFusion error

OwnedTableReference

Represents a path to a table that may require further resolution that owns the underlying names

SchemaError

Schema-related errors

TableReference

Represents a path to a table that may require further resolution

Traits

ExprSchema

Provides schema information needed by certain methods of Expr (defined in the datafusion-common crate).

ToDFSchema

Convenience trait to convert Schema like things to DFSchema and DFSchemaRef with fewer keystrokes

Functions

field_not_found

Create a “field not found” DataFusion::SchemaError

reverse_sort_options

Computes the “reverse” of given SortOptions.

Type Definitions

DFSchemaRef

A reference-counted reference to a DFSchema.

Result

Result type for operations that could result in an DataFusionError

SharedResult

Result type for operations that could result in an DataFusionError and needs to be shared (wrapped into Arc).