Struct datafusion_common::Column

pub struct Column {
    pub relation: Option<String>,
    pub name: String,
}

A named reference to a qualified field in a schema.

Fields
relation: Option<String>

relation/table name.

name: String

field/column name.

Implementations

impl Column

pub fn new(relation: Option<impl Into<String>>, name: impl Into<String>) -> Self

Create Column from optional qualifier and name

pub fn from_name(name: impl Into<String>) -> Self

Create Column from unqualified name.

pub fn from_qualified_name(flat_name: impl Into<String>) -> Self

Deserialize a fully qualified name string into a column

pub fn flat_name(&self) -> String

Serialize column into a flat name string

pub fn normalize_with_schemas( self, schemas: &[&Arc<DFSchema>], using_columns: &[HashSet<Column>] ) -> Result<Self>

Qualify column if not done yet.

If this column already has a relation, it will be returned as is and the given parameters are ignored. Otherwise this will search through the given schemas to find the column. This will use the first schema that matches.

A schema matches if there is a single column that – when unqualified – matches this column. There is an exception for USING statements, see below.

Using columns

Take the following SQL statement:

SELECT id FROM t1 JOIN t2 USING(id)

In this case, both t1.id and t2.id will match unqualified column id. To express this possibility, use using_columns. Each entry in this array is a set of columns that are bound together via a USING clause. So in this example this would be [{t1.id, t2.id}].