Struct datafusion_common::DFSchema

pub struct DFSchema {
    pub fields: Vec<DFField>,
    pub metadata: HashMap<String, String>,
}

DFSchema wraps an Arrow schema and adds relation names

Fields
fields: Vec<DFField>

Fields

metadata: HashMap<String, String>

Additional metadata in form of key value pairs

Implementations

impl DFSchema

pub fn empty() -> Self

Creates an empty DFSchema

pub fn new(fields: Vec<DFField>) -> Result<Self>

๐Deprecated since 7.0.0: please use new_with_metadata instead

Create a new DFSchema

pub fn new_with_metadata( fields: Vec<DFField>, metadata: HashMap<String, String> ) -> Result<Self>

Create a new DFSchema

pub fn try_from_qualified_schema( qualifier: &str, schema: &Schema ) -> Result<Self>

Create a DFSchema from an Arrow schema

pub fn join(&self, schema: &DFSchema) -> Result<Self>

Create a new schema that contains the fields from this schema followed by the fields from the supplied schema. An error will be returned if there are duplicate field names.

pub fn merge(&mut self, other_schema: &DFSchema)

Modify this schema by appending the fields from the supplied schema, ignoring any duplicate fields.

pub fn fields(&self) -> &Vec<DFField>

Get a list of fields

pub fn field(&self, i: usize) -> &DFField

Returns an immutable reference of a specific Field instance selected using an offset within the internal fields vector

pub fn index_of(&self, name: &str) -> Result<usize>

๐Deprecated since 8.0.0: please use index_of_column_by_name instead

Find the index of the column with the given unqualified name

pub fn index_of_column_by_name( &self, qualifier: Option<&str>, name: &str ) -> Result<usize>

pub fn index_of_column(&self, col: &Column) -> Result<usize>

Find the index of the column with the given qualifier and name

pub fn field_with_name( &self, qualifier: Option<&str>, name: &str ) -> Result<&DFField>

Find the field with the given name

pub fn fields_with_qualified(&self, qualifier: &str) -> Vec<&DFField>

Find all fields having the given qualifier

pub fn fields_with_unqualified_name(&self, name: &str) -> Vec<&DFField>

Find all fields match the given name

pub fn field_with_unqualified_name(&self, name: &str) -> Result<&DFField>

Find the field with the given name

pub fn field_with_qualified_name( &self, qualifier: &str, name: &str ) -> Result<&DFField>

Find the field with the given qualified name

pub fn field_from_column(&self, column: &Column) -> Result<&DFField>

Find the field with the given qualified column

pub fn matches_arrow_schema(&self, arrow_schema: &Schema) -> bool

Check to see if unqualified field names matches field names in Arrow schema

pub fn check_arrow_schema_type_compatible( &self, arrow_schema: &Schema ) -> Result<()>

Check to see if fields in 2 Arrow schemas are compatible

pub fn equivalent_names_and_types(&self, other: &Self) -> bool

Returns true if the two schemas have the same qualified named fields with the same data types. Returns false otherwise.

This is a specialized version of Eq that ignores differences in nullability and metadata.

pub fn strip_qualifiers(self) -> Self

Strip all field qualifier in schema

pub fn replace_qualifier(self, qualifier: &str) -> Self

Replace all field qualifier with new value in schema

pub fn field_names(&self) -> Vec<String>

Get list of fully-qualified field names in this schema

pub fn metadata(&self) -> &HashMap<String, String>

Get metadata of this schema